Add functionality for Password Encryption
Added a hashing algorithm for our passwords so that any correct password will be put into a hash table with a random key. On clicking of the logout button the hash table will be cleared of all contents. So only one correct password will be present in the hash table at a time.